moved UsbConfigState interface to stores
