+++ ui/src/utils.ts
@@ -232,11 +232,3 @@ export function isChromeOS() {
   /* ChromeOS sets navigator.platform to Linux :/ */
   return !!navigator.userAgent.match(" CrOS ");
 }
-
-export interface UsbConfigState {
-  vendor_id: string;
-  product_id: string;
-  serial_number: string;
-  manufacturer: string;
-  product: string;
-}
